api: initial removal of the jax-rs dependencies
In order to remove the jax-rs dependencies (WebTarget, MediaType,
Entity, etc...) the new OpenStackTarget class has been introduced.
The OpenStackTarget class objective is to abstract the get, post,
put, delete requests so that it will be possible to use different
versions of the jax-rs specification.
In those cases where removing such dependencies was too complex
(e.g. uploading/downloading an image in glace, the swift module,
etc...) the OpenStackTarget.getWebTarget method has been used in
order to retrieve the WebTarget object and maintain the previous
implementation. In the future also these corner cases will be
abstracted and handled in the proper way.

Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>